clean repo
