+++ exec_files/get_cmds.c
@@ -6,7 +6,7 @@
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/25 17:11:00 by achane-l          #+#    #+#             */
-/*   Updated: 2022/02/28 00:18:55 by achane-l         ###   ########.fr       */
+/*   Updated: 2022/02/28 17:52:36 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
