+++ utils/ft_strchr.c
@@ -1,41 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_echo.c                                          :+:      :+:    :+:   */
+/*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/02/28 14:56:15 by achane-l          #+#    #+#             */
-/*   Updated: 2022/02/28 17:42:53 by achane-l         ###   ########.fr       */
+/*   Created: 2022/02/28 17:44:15 by achane-l          #+#    #+#             */
+/*   Updated: 2022/02/28 17:44:50 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "./built-in.h"
+#include "./utils.h"
 
-int	ft_echo(char **args)
+char	*ft_strchr(const char *s, int c)
 {
-	int	count_args;
-	int i;
-	bool	nl;
+	char	*str;
 
-	count_args = 0;
-	i = 1;
-	nl = true;
-	while(args[count_args])
-		count_args++;
-	while(i < count_args)
+	str = (char *)s;
+	while (*str != c)
 	{
-		if (i == 1 && args[i] && strcmp(args[i], "-n") == 0)
+		if (*str == '\0')
 		{
-			i++;
-			nl = false;
+			return (NULL);
 		}
-		write(1, args[i], ft_strlen(args[i]));
-		if (i != count_args - 1)
-			write(1, " ", 1);
-		i++;
+		str++;
 	}
-	if (nl)
-		write(1, "\n", 1);
-	return (1);
+	return (str);
 }
