+++ utils/ft_strcmp.c
@@ -1,41 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_echo.c                                          :+:      :+:    :+:   */
+/*   ft_strcmp.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: achane-l <achane-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/02/28 14:56:15 by achane-l          #+#    #+#             */
-/*   Updated: 2022/02/28 17:42:53 by achane-l         ###   ########.fr       */
+/*   Created: 2022/02/28 17:43:45 by achane-l          #+#    #+#             */
+/*   Updated: 2022/02/28 17:43:58 by achane-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "./built-in.h"
+#include "./utils.h"
 
-int	ft_echo(char **args)
+int	ft_strcmp(char *s1, char *s2)
 {
-	int	count_args;
 	int i;
-	bool	nl;
 
-	count_args = 0;
-	i = 1;
-	nl = true;
-	while(args[count_args])
-		count_args++;
-	while(i < count_args)
+	i = 0;
+	while (s1[i] != '\0' || s2[i] != '\0')
 	{
-		if (i == 1 && args[i] && strcmp(args[i], "-n") == 0)
-		{
+		if (s1[i] != s2[i])
+			return (s1[i] - s2[i]);
+		else
 			i++;
-			nl = false;
-		}
-		write(1, args[i], ft_strlen(args[i]));
-		if (i != count_args - 1)
-			write(1, " ", 1);
-		i++;
 	}
-	if (nl)
-		write(1, "\n", 1);
-	return (1);
+	return (0);
 }
